Correct game lobby manager initGameConfig
