+++ core/src/main/java/xiao/battleroyale/common/message/zone/ZoneMessageManager.java
@@ -82,7 +82,7 @@ public class ZoneMessageManager extends AbstractMessageManager<ZoneMessage> impl
     @Override
     public void initGameConfig(ServerLevel serverLevel) {}
     @Override
-    public boolean isPreparedForGame() { return true; }
+    public boolean isConfigPrepared() { return true; }
     @Override
     public void initGame(ServerLevel serverLevel) {}
     @Override
